backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Security-Too==5.5.2
Flask-CORS==5.0.1
Flask-Caching==2.3.1
Flask-Mail==0.10.0
SQLAlchemy==2.0.36
Werkzeug==3.1.3
celery[redis]==5.4.0
redis==5.2.1
bcrypt==4.2.1
python-dotenv==1.0.1
WeasyPrint==62.3
PyPDF2==3.0.1
flask-swagger-ui==5.21.0
pyyaml==6.0.2
setuptools<81
